+++ app/src/main/java/com/boolenull/sensortest/model/EnumMySensor.kt
