+++ app/src/main/java/com/boolenull/sensortest/ui/fragment/LightFragment.kt
@@ -47,10 +47,12 @@ class LightFragment: Fragment(), MySensorEventListener {
 
     override fun onSensorChanged(event: SensorEvent?) {
         if (points.size > maxPoint) points.clear()
-        view!!.tv.text = getString(R.string.light, event!!.values[0])
-        points.add(DataPoint(points.size.toDouble(), event.values[0].toDouble()))
-        val series = LineGraphSeries<DataPoint>(points.toTypedArray())
-        graph.removeAllSeries()
-        graph.addSeries(series)
+        event?.let {
+            view!!.tv.text = getString(R.string.light, it.values[0])
+            points.add(DataPoint(points.size.toDouble(), it.values[0].toDouble()))
+            val series = LineGraphSeries<DataPoint>(points.toTypedArray())
+            graph.removeAllSeries()
+            graph.addSeries(series)
+        }
     }
 }
