+++ app/src/main/java/com/boolenull/sensortest/ui/fragment/TempFragment.kt
@@ -48,10 +48,12 @@ class TempFragment: Fragment(), MySensorEventListener {
 
     override fun onSensorChanged(event: SensorEvent?) {
         if (points.size > maxPoint) points.clear()
-        view!!.tv.text = getString(R.string.temp) + " " + event!!.values[0]
-        points.add(DataPoint(points.size.toDouble(), event.values[0].toDouble()))
-        val series = LineGraphSeries<DataPoint>(points.toTypedArray())
-        graph.removeAllSeries()
-        graph.addSeries(series)
+        event?.let {
+            view!!.tv.text = getString(R.string.temp, it.values[0])
+            points.add(DataPoint(points.size.toDouble(), it.values[0].toDouble()))
+            val series = LineGraphSeries<DataPoint>(points.toTypedArray())
+            graph.removeAllSeries()
+            graph.addSeries(series)
+        }
     }
 }
